requirements.txt
flask==2.2.3
pyyaml==6.0
requests==2.28.2
werkzeug==2.2.3
python-dotenv==1.0.0
importlib-metadata==6.1.0
